Move fragments and scenes into different directories
