+++ screen-extractor/kartalytics/kartalytics.rb
@@ -1,0 +1,4 @@
+require_relative 'scene/final_result'
+
+module Kartalytics
+end
